Yuchenyeo2023/is4 32 upload additional images and show all uploaded images

+++ backend/src/main/java/IS442/G1T3/IDPhotoGenerator/dto/CropRequestDTO.java
@@ -6,9 +6,9 @@ import lombok.Setter;
 @Getter
 @Setter
 public class CropRequestDTO {
-    private double x;
-    private double y;
-    private double width;
-    private double height;
+    private int x;
+    private int y;
+    private int width;
+    private int height;
 
 }
